Split geocode controller from maps controller
